requirements.txt
Flask
Twilio
webrtcvad
requests
langchain_google_genai
langchain_core
python-dotenv
